feat: implement NutritionistPatient management with command and query services

+++ services/nutritionist-service/src/main/java/pe/edu/upc/nutritionist_service/nutritionist/infrastructure/persistence/jpa/repositories/NutritionistRepository.java
@@ -7,6 +7,6 @@ import pe.edu.upc.nutritionist_service.nutritionist.domain.model.aggregates.Nutr
 import java.util.Optional;
 
 @Repository
-public interface NutritionistRepository extends JpaRepository<Nutritionist, Integer> {
+public interface NutritionistRepository extends JpaRepository<Nutritionist, Long> {
     Optional<Nutritionist> findByUserId(Long userId);
 }
